💫 UPDATE: move InitialBlockSelector to a separate file
